+++ packages/wpcom.js/test/test.wpcom.site.media.js
@@ -1,46 +1,50 @@
 /**
  * Module dependencies
  */
-var util = require( './util' );
-var assert = require( 'assert' );
+const util = require( './util' );
+const assert = require( 'assert' );
 
 /**
  * Testing data
  */
-var fixture = require( './fixture' );
+const fixture = require( './fixture' );
 
 describe( 'wpcom.site.media', function() {
 	// Global instances
-	var wpcom = util.wpcom();
-	var site = wpcom.site( util.site() );
-	var add_urls_array;
-	var add_urls_object;
+	const wpcom = util.wpcom();
+	const site = wpcom.site( util.site() );
+	let add_urls_array;
+	let add_urls_object;
 
 	// Create a testing_media before to start tests
 
-	var testing_media;
+	let testing_media;
 	before( done => {
-		site.addMediaFiles( fixture.media.files[1] )
+		site.addMediaFiles( fixture.media.files[ 1 ] )
 			.then( data => {
-				testing_media = data ? data.media[0] : {};
+				testing_media = data ? data.media[ 0 ] : {};
 				done();
 			} )
 			.catch( done );
 	} );
 
 	after( done => {
+		if ( ! add_urls_array || add_urls_array.length < 1 ) {
+			return done();
+		}
+
 		// clean media added through of array by urls
-		site.deleteMedia( add_urls_array.media[0].ID )
-			.then( () => site.deleteMedia( add_urls_array.media[1].ID ) )
-			.then( () => site.deleteMedia( add_urls_array.media[2].ID ) )
-			.then( () => site.deleteMedia( add_urls_object.media[0].ID ) )
+		site.deleteMedia( add_urls_array.media[ 0 ].ID )
+			.then( () => site.deleteMedia( add_urls_array.media[ 1 ].ID ) )
+			.then( () => site.deleteMedia( add_urls_array.media[ 2 ].ID ) )
+			.then( () => site.deleteMedia( add_urls_object.media[ 0 ].ID ) )
 			.then( () => done() )
 			.catch( done );
 	} );
 
 	describe( 'wpcom.site.media.get', function() {
 		it( 'should get added media', done => {
-			var media = site.media( testing_media.ID );
+			const media = site.media( testing_media.ID );
 			media.get()
 				.then( data => {
 					assert.equal( testing_media.ID, data.ID );
@@ -52,10 +56,11 @@ describe( 'wpcom.site.media', function() {
 
 	describe( 'wpcom.site.media.update', function() {
 		it( 'should edit the media title', done => {
-			var edited_title = 'This is the new title';
+			const edited_title = 'This is the new title';
 
-			site.media( testing_media.ID )
-			.update( { apiVersion: '1.1' }, { title: edited_title } )
+			site
+				.media( testing_media.ID )
+				.update( { apiVersion: '1.1' }, { title: edited_title } )
 				.then( data => {
 					assert.ok( data );
 					assert.equal( edited_title, data.title );
@@ -81,7 +86,7 @@ describe( 'wpcom.site.media', function() {
 
 	describe( 'wpcom.site.media.addUrls', function() {
 		it( 'should create a new media from an object', done => {
-			var media_object = fixture.media.urls[1];
+			const media_object = fixture.media.urls[ 1 ];
 
 			site.media().addUrls( media_object )
 				.then( data => {